Added common function for ECS resource tags
